Wired in the validators

+++ models/comments.model.js
@@ -1,10 +1,20 @@
 const logger = require('../logger/logger.js');
 const db = require('../db/connection.js');
+const miscService = require('../services/misc.service.js');
+const miscValidator = require('../validators/misc.validator.js');
+const commentValidator = require('../validators/comment.validator.js');
+const userValidator = require('../validators/user.validator.js');
 
 const selectCommentsByArticleId = (articleId, limit = 10, page = 1) => {
   logger.debug(`In selectCommentsByArticleId() in comments.model`);
   logger.info(`Selecting comments from database where article_id:${articleId} `
       + `limit=${limit} page=${page}`);
+
+  // Validate article ID
+  const idValObj = miscValidator.validateId(articleId);
+  if (!idValObj.valid) {
+    return Promise.reject({status: 400, msg: idValObj.msg});
+  }
 
   const queryStr = 
       `SELECT * FROM comments 
@@ -12,7 +22,10 @@ const selectCommentsByArticleId = (articleId, limit = 10, page = 1) => {
         ORDER BY created_at DESC 
         LIMIT $2 OFFSET $3;`; 
 
-  return db.query(queryStr, [articleId, limit, (page - 1) * limit])
+  return miscService.checkValueExists('articles', 'article_id', articleId)
+      .then(() => {
+        return db.query(queryStr, [articleId, limit, (page - 1) * limit]);
+      })
       .then(({rows: comments}) => {
         return comments;
       });
@@ -23,12 +36,41 @@ const createComment = (articleId, author, body) => {
   logger.info(`Inserting comment into database where article_id:${articleId} `
       + `author:${author}`);
 
-  return db
-      .query(
-          `INSERT INTO comments(article_id, author, body) 
-            VALUES ($1, $2, $3)
-            RETURNING *;`,
-          [articleId, author, body])
+  // Validate article ID
+  const idValObj = miscValidator.validateId(articleId);
+  if (!idValObj.valid) {
+    return Promise.reject({status: 400, msg: idValObj.msg});
+  }
+  
+  // Validate author
+  const userValObj = userValidator.validateUsername(author);
+  if (!userValObj.valid) {
+    return Promise.reject({status: 400, msg: userValObj.msg});
+  }
+
+  // Validate body
+  const bodyValObj = commentValidator.validateBody(body);
+  if (!bodyValObj.valid) {
+    return Promise.reject({status: 400, msg: bodyValObj});
+  }
+
+  // Check article ID exists
+  const artProm = miscService.checkValueExists('articles', 'article_id', articleId);
+  // Check username exists
+  const userProm = miscService.checkValueExists('users', 'username', author);
+
+  const promArr = [];
+  promArr.push(artProm);
+  promArr.push(userProm);
+
+  return Promise.all(promArr)
+      .then(() => {
+        return db.query(
+            `INSERT INTO comments(article_id, author, body) 
+              VALUES ($1, $2, $3)
+              RETURNING *;`,
+            [articleId, author, body])
+      })
       .then(({rows}) => {
         return rows[0];
       });
@@ -37,6 +79,12 @@ const createComment = (articleId, author, body) => {
 const deleteCommentById = (commentId) => {
   logger.debug(`In deleteCommentById() in comments.model`);
   logger.info(`Deleting comment from database where comment_id:${commentId}`);
+
+  // Validate ID
+  const idValObj = miscValidator.validateId(commentId);
+  if (!idValObj.valid) {
+    return Promise.reject({status: 400, msg: idValObj.msg});
+  }  
 
   return db
       .query(`DELETE FROM comments WHERE comment_id = $1 RETURNING *;`,
@@ -56,6 +104,18 @@ const increaseVoteByCommentId = (commentId, incVotes) => {
   logger.debug(`Changing vote count on comment where comment_id:${commentId}`
       + ` with vote_increment:${incVotes}`);
   
+  // Validate ID
+  const idValObj = miscValidator.validateId(commentId);
+  if (!idValObj.valid) {
+    return Promise.reject({status: 400, msg: idValObj.msg});
+  }
+
+  // Validate votes
+  const voteValObj = commentValidator.validateVote(incVotes);
+  if (!voteValObj.valid) {
+    return Promise.reject({status: 400, msg: voteValObj.msg});
+  }
+
   return db
       .query(
           `UPDATE comments
